Implemented multi version for Magento projects when setting up docker
Made changes to support M2.3.3/2.3.7.

bin/setup supports a second parameter -v (acccepts 2.3.3, 2.3.7 or latest as values).
Based on the parameter, the proper docker images will be used. M2.3.3 has a problem with images, and we need to setup the images for 2.3.7 in order to run composer install, and then switch the images for 2.3.3.
bin/n98 installs the correct version for the project we setup.